% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{AddLinkCorrelations}
\alias{AddLinkCorrelations}
\title{Add correlations between region accessiblity and gene expression to region-gene links}
\usage{
AddLinkCorrelations(df, col1, col2, mat1, mat2, n.cores = 1)
}
\arguments{
\item{df}{Dataframe linking two types of features (i.e. TFs & genomic regions, genomic regions & genes, etc)}

\item{col1}{Column name for the 1st feature (i.e. "region")}

\item{col2}{Column name for the 2nd feature (i.e. "gene")}

\item{mat1}{Matrix corresponding to feature 1. Columns must match columns of mat2.}

\item{mat2}{Matrix corresponding to feature 2. Columns must match columns of mat1}

\item{n.cores}{Number of cores for parallel processing}
}
\value{
Dataframe with the "cor" column added
}
\description{
Add correlations between region accessiblity and gene expression to region-gene links
}
