% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{AddRegionGeneCor}
\alias{AddRegionGeneCor}
\title{Add correlations between region accessiblity and gene expression to region-gene links}
\usage{
AddRegionGeneCor(df, region.mat, rna.mat)
}
\arguments{
\item{df}{Dataframe from RegionGeneLinks Must contain "region" and "gene" columns.}

\item{region.mat}{Matrix of region accessibility. Column names must match rna.mat column names}

\item{rna.mat}{Matrix of gene expression. Column names must match region.mat column names.}
}
\value{
Dataframe with the "cor" column added
}
\description{
Add correlations between region accessiblity and gene expression to region-gene links
}
