% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{buildRegionMatrix}
\alias{buildRegionMatrix}
\title{Get an approximate accessibility matrix for arbitrary regions by using the peaks overlapping those regions}
\usage{
buildRegionMatrix(regions.gr, peaks.gr, peaks.mat)
}
\arguments{
\item{regions.gr}{GRanges object of regions to create matrix for}

\item{peaks.gr}{GRanges object of peaks in peaks matrix}

\item{peaks.mat}{Sparse peak accessibility matrix. Rows must match peaks.gr}
}
\value{
A matrix of accessibility for each region
}
\description{
Get an approximate accessibility matrix for arbitrary regions by using the peaks overlapping those regions
}
