% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_access.R
\name{CalcDiffAccess}
\alias{CalcDiffAccess}
\title{Compute differentially accessible peaks for each cluster}
\usage{
CalcDiffAccess(counts, clusters, min.p = 1e-100)
}
\arguments{
\item{counts}{Binary accessibility matrix (peaks x cells)}

\item{clusters}{Cluster assignments encoded as a factor}

\item{min.p}{Minimum p-value possible}
}
\value{
List of dataframes with cluster specificity, logfc, fraction accessible,
        deviation from expected accessibility, and p-value for each site tested
}
\description{
Compute differentially accessible peaks for each cluster
}
