% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_enrich.R
\name{ChromRegionEnrich}
\alias{ChromRegionEnrich}
\title{Compute enrichment in each column of the chromatin accessibility matrix for all regions in regionsList}
\usage{
ChromRegionEnrich(SE, regionsList, n.runs = 5)
}
\arguments{
\item{SE}{Peaks accessibility matrix in SummarizedExperiment format}

\item{regionsList}{List of regions in GenomicRanges format}

\item{n.runs}{Number of replicates to run}
}
\value{
Matrix of z-scores representing the enrichment of each set of regions in each column of SE
}
\description{
Compute enrichment in each column of the chromatin accessibility matrix for all regions in regionsList
}
