% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{cluster_pheno_pvals}
\alias{cluster_pheno_pvals}
\title{Calculate p-values linking phenotypes to clusters using permutation testing}
\usage{
cluster_pheno_pvals(
  pheno.blocks,
  top_diff_peaks,
  background_peaks,
  n.rand = 10000,
  n.cores = 8,
  z.score = T
)
}
\arguments{
\item{pheno.blocks}{A list of genomic regions for each GWAS phenotype in granges format}

\item{top_diff_peaks}{A list of top accessible sites for each cluster in granges format}

\item{background_peaks}{A granges of background accessible peaks to sample from}

\item{n.rand}{Number of permutations}

\item{n.cores}{Number of cores for parallel processing}

\item{score.col}{Whether or not to use z-scores or just the number of overlapping regions}
}
\value{
A matrix of empirical p-values for each phenotype/cluster enrichment test
}
\description{
Calculate p-values linking phenotypes to clusters using permutation testing
}
