% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{cluster_pheno_scores}
\alias{cluster_pheno_scores}
\title{Intersect phenotype-linked LD blocks with peaks}
\usage{
cluster_pheno_scores(pheno.blocks, top_diff_peaks, z.score = T)
}
\arguments{
\item{pheno.blocks}{A list of genomic regions for each GWAS phenotype in granges format}

\item{top_diff_peaks}{A list of top accessible sites for each cluster in granges format}

\item{z.score}{Whether or not to use z-scores or just the number of overlapping regions}
}
\value{
A matrix of overlap scores for each phenotype/cluster combination
}
\description{
Intersect phenotype-linked LD blocks with peaks
}
