% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extend_gr}
\alias{extend_gr}
\title{Extend granges by buffer length}
\usage{
extend_gr(gr, buffer, chr.lengths)
}
\arguments{
\item{gr}{GRanges object}

\item{buffer}{bp to extend gr by}

\item{chr.lengths}{chromosome lengths}
}
\value{
GRanges object with each range extended by the buffer
}
\description{
Extend granges by buffer length
}
