% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{extract_sig_mutations}
\alias{extract_sig_mutations}
\title{Extracts significant mutations for each cluster}
\usage{
extract_sig_mutations(pheno.blocks, top_diff_peaks)
}
\arguments{
\item{pheno.blocks}{A list of genomic regions for each GWAS phenotype in granges format}

\item{top_diff_peaks}{A list of top accessible sites for each cluster in granges format}
}
\value{
A list of lists for each phenotype/cluster combination with
}
\description{
Extracts significant mutations for each cluster
}
