% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extractField}
\alias{extractField}
\title{Extract the desired field from a delimited string}
\usage{
extractField(string, field = 1, delim = "_")
}
\arguments{
\item{string}{Input string}

\item{field}{Field position}

\item{delim}{Delimiter (i.e. tabs, spaces, underscores, commas)}
}
\value{
Extracted field
}
\description{
Extract the desired field from a delimited string
}
