% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{filter_snps}
\alias{filter_snps}
\title{Filter SNPs by p-value}
\usage{
filter_snps(pheno.snp.list, p.cutoff, max.snps = Inf, min.snps = 1)
}
\arguments{
\item{pheno.snp.list}{List of z-scores for the SNPs in each phenotype}

\item{p.cutoff}{P-value cutoff}

\item{max.snps}{Max number of SNPs for each phenotype}

\item{min.snps}{Phenotypes with fewer than this number of snps passing filter will be removed}
}
\value{
List of filtered z-scores for the SNPs in each phenotype
}
\description{
Filter SNPs by p-value
}
