% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_access.R
\name{find_all_diff-deprecated}
\alias{find_all_diff-deprecated}
\title{find_all_diff}
\arguments{
\item{counts}{Binary accessibility matrix (peaks x cells)}

\item{clusters}{Cluster assignments encoded as a factor}

\item{min.p}{Minimum p-value possible}
}
\value{
List of dataframes with cluster specificity, logfc, and p-value for each site tested
}
\description{
Find differentially accessible peaks/genes for each cluster
}
\seealso{
\code{\link{chromfunks-deprecated}}
}
\keyword{internal}
