% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_genes.R
\name{FindCausalPeaks}
\alias{FindCausalPeaks}
\title{Identify most impactful SNPs for a given trait/cell type combo}
\usage{
FindCausalPeaks(SE, traits, bg.peaks = NULL, min.z = 1, min.PP = 0.025)
}
\arguments{
\item{SE}{Accessibility counts as a Summarized Experiment object}

\item{traits}{Posterior probability score of each peak for every trait as a Summarized Experiment object}

\item{bg.peaks}{Matrix of background peaks. Will be computed with getBackgroundPeaks if not specified}

\item{min.z}{Z-score cutoff for peaks}

\item{min.PP}{Posterior probability score for peaks}
}
\value{
List of lists of causal peaks for each trait and cell type. Each trait/cell type pair has a dataframe with the z-score, accessibility deviation, and posterior probability for each causal peak
}
\description{
Identify most impactful SNPs for a given trait/cell type combo
}
