% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_genes.R
\name{FindCausalProb}
\alias{FindCausalProb}
\title{Find the posterior probability score of causal peaks coaccessible with genes for specified trait/cluster pairs}
\usage{
FindCausalProb(pheno.genes.mat, causal.z, peak.gene.weights)
}
\arguments{
\item{pheno.genes.mat}{Matrix describing the coaccessibility of specified trait/cluster pairs with genes. Generated by FindLinkedGenes.}

\item{causal.z}{List of lists of causal peaks for each trait and cell type generated by FindCausalPeaks}

\item{peak.gene.weights}{Matrix of peak - gene coaccessibilities}
}
\value{
Matrix of the sum of posterior probabilities of causal snps driving the coaccessibility of specified trait/cluster pairs with genes
}
\description{
Find the posterior probability score of causal peaks coaccessible with genes for specified trait/cluster pairs
}
