% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_genes.R
\name{FindLinkedGenes}
\alias{FindLinkedGenes}
\title{Find genes coaccessible with causal peaks for each cluster/trait pair}
\usage{
FindLinkedGenes(pheno.cl.use, causal.z, peak.gene.weights, min.weight = 0.1)
}
\arguments{
\item{pheno.cl.use}{Trait/cluster pairs to analyze, as a character vector of "trait: cluster" elements}

\item{causal.z}{List of lists of causal peaks for each trait and cell type generated by FindCausalPeaks}

\item{peak.gene.weights}{Matrix of peak - gene coaccessibilities}

\item{min.weight}{Minimum coaccessibility to consider}
}
\value{
Matrix describing the coaccessibility of specified trait/cluster pairs with genes
}
\description{
Find genes coaccessible with causal peaks for each cluster/trait pair
}
