% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{FindNearbyGenes}
\alias{FindNearbyGenes}
\title{Find all genes within specified distance of each genomic region}
\usage{
FindNearbyGenes(regions, genes.anno, distance, chr.lengths, region.name = NULL)
}
\arguments{
\item{genes.anno}{Gene annotations generated by the geneAnno function}

\item{distance}{Distance to search for genes}

\item{chr.lengths}{Chromosome sizes}

\item{region.name}{Name of each region if in the GRanges metadata. Otherwise set to null.}

\item{gr}{Genomic regions}
}
\value{
List of genes within specified distance of genomic regions
}
\description{
Find all genes within specified distance of each genomic region
}
