% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{flattenMat}
\alias{flattenMat}
\title{Flatten Matrix into a vector}
\usage{
flattenMat(mat, delim = "_")
}
\arguments{
\item{mat}{Matrix}

\item{delim}{Delimiter to separate row and column names}
}
\value{
Vector with row and column names concatenated by delim
}
\description{
Flatten Matrix into a vector
}
