% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{formatCisTopicPeaks}
\alias{formatCisTopicPeaks}
\title{Reformat rownames}
\usage{
formatCisTopicPeaks(peak.names)
}
\arguments{
\item{peak.names}{Peak names where the chromosomes and base pairs are separated by underscores (chr_start_end)}
}
\value{
Peak names formatted as so: chr:start-end
}
\description{
Reformat rownames
}
