% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{geneAnno}
\alias{geneAnno}
\title{Get gene body information from a human txdb object}
\usage{
geneAnno(txdb)
}
\arguments{
\item{txdb}{TxDb object (i.e. TxDb.Hsapiens.UCSC.hg38.knownGene)}
}
\value{
GRanges of gene body annotations
}
\description{
Get gene body information from a human txdb object
}
