% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{get_coaccess_peaks}
\alias{get_coaccess_peaks}
\title{Reformat coaccessible peaks in list format}
\usage{
get_coaccess_peaks(peaks, conns)
}
\arguments{
\item{peaks}{to find coaccessible peaks for}

\item{conns}{Dataframe of peak to peak connections}
}
\value{
List of coaccessible peaks (in granges format) for each peak
}
\description{
Reformat coaccessible peaks in list format
}
