% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{get_pheno_blocks}
\alias{get_pheno_blocks}
\title{Compute blocks for each GWAS phenotype by finding the tag snp for in each LD block and extending by the buffer size}
\usage{
get_pheno_blocks(
  sig.pheno.snps,
  snp.data,
  ld_blocks,
  chr.lengths,
  buffer.size = 20000,
  max.block.size = Inf,
  max.blocks = Inf
)
}
\arguments{
\item{sig.pheno.snps}{List of SNPs for each GWAS phenotype that pass the p-value threshold}

\item{snp.data}{Dataframe of SNP genomic location and other info generated by get_snp_data}

\item{ld_blocks}{Granges of LD blocks}

\item{chr.lengths}{Dataframe of chromosome lengths}

\item{buffer.size}{Buffer size around each tag snp}

\item{max.block.size}{Maximum size of each block (larger blocks will be split)}

\item{max.blocks}{Maximum number of LD blocks for each phenotype}
}
\value{
List of blocks for each phenotype defined by the top tag snp for that block
}
\description{
Compute blocks for each GWAS phenotype by finding the tag snp for in each LD block and extending by the buffer size
}
