% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getBulkDev}
\alias{getBulkDev}
\title{Compute deviations on a ATAC counts matrix}
\usage{
getBulkDev(counts)
}
\arguments{
\item{counts}{Bulk or pseudo-bulk ATAC counts matrix}
}
\value{
A matrix of deviations from the expected counts
}
\description{
Compute deviations on a ATAC counts matrix
}
