% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.R
\name{GetClusterAvg}
\alias{GetClusterAvg}
\title{Average counts matrix by cluster/cell type}
\usage{
GetClusterAvg(counts, idents, binarize = F)
}
\arguments{
\item{counts}{Counts matrix}

\item{idents}{Cluster/Cell Type identities}

\item{binarize}{Binarize the matrix before computing averages}
}
\value{
Counts matrix averaged over clusters/cell types
}
\description{
Average counts matrix by cluster/cell type
}
