% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getPseudobulk}
\alias{getPseudobulk}
\title{Helper function for generating a pseudo-bulk matrix}
\usage{
getPseudobulk(mat, celltype)
}
\arguments{
\item{mat}{Input single cell matrix}

\item{celltype}{Cell type anotations}
}
\value{
A pseudo-bulk matrix where the columns have been summed by cell type
}
\description{
Helper function for generating a pseudo-bulk matrix
}
