% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.R
\name{GetRegionCounts}
\alias{GetRegionCounts}
\title{Build accessibility matrices for peaks overlapping list of arbitrary regions}
\usage{
GetRegionCounts(peak.counts, peak.maps, idents = NULL, scale = F)
}
\arguments{
\item{peak.counts}{Chromatin accessibility counts matrix}

\item{peak.maps}{List of peaks}

\item{idents}{Cell identities (defaults to NULL)}

\item{scale}{Whether or not to scale the single cell matrices}
}
\value{
List of counts matrices
}
\description{
Build accessibility matrices for peaks overlapping list of arbitrary regions
}
