% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{granges2df}
\alias{granges2df}
\title{Convert gRanges to dataframe}
\usage{
granges2df(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
GRanges info in dataframe format
}
\description{
Convert gRanges to dataframe
}
