% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{granges2peak}
\alias{granges2peak}
\title{Function for converting a granges object into a character vector of peak names}
\usage{
granges2peak(gr, delim = c(":", "-"))
}
\arguments{
\item{gr}{GenomicRanges object}

\item{delim}{Peak name delimiter}
}
\value{
Character vector of peak names
}
\description{
Function for converting a granges object into a character vector of peak names
}
