% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_genes.R
\name{ImportCausalSNPs}
\alias{ImportCausalSNPs}
\title{Import causal snp bed files}
\usage{
ImportCausalSNPs(files, colidx = 5)
}
\arguments{
\item{files}{Paths of bed files to import}

\item{colidx}{Index of the posterior probability column}
}
\value{
Causal SNPs as GRanges objects
}
\description{
Import causal snp bed files
}
