% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{invertListVector}
\alias{invertListVector}
\title{Inverts a list of named numeric vectors}
\usage{
invertListVector(my.list)
}
\arguments{
\item{my.list}{List}
}
\value{
Inverted list
}
\description{
Inverts a list of named numeric vectors
}
