% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.R
\name{LogTFIDF}
\alias{LogTFIDF}
\title{Compute the log TF-IDF transform for accessibility matrices
Adapted from Signac}
\usage{
LogTFIDF(counts, scale.factor = 10000, verbose = T)
}
\arguments{
\item{counts}{Accessibility counts matrix}

\item{scale.factor}{Factor to scale normalized counts with. Default is 10000.}

\item{verbose}{Print progress}
}
\value{
Counts matrix normalized with the log(TF-IDF + 1) transform
}
\description{
Compute the log TF-IDF transform for accessibility matrices
Adapted from Signac
}
