% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{MotifEnrich}
\alias{MotifEnrich}
\title{Computes motif enrichment with a hypergeometric distribution}
\usage{
MotifEnrich(
  meta.features,
  features,
  motif.all,
  n.background = 40000,
  verbose = TRUE
)
}
\arguments{
\item{meta.features}{Dataframe with gc content and peak names}

\item{features}{Vector peak names to test for enrichment. All peaks must be in meta.features}

\item{motif.all}{Matrix of peaks by motifs}

\item{n.background}{Number of background peaks to sample}

\item{verbose}{Print progress}
}
\value{
Dataframe with motif enrichment results
}
\description{
Computes motif enrichment with a hypergeometric distribution
}
