% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{MotifRegionLinks}
\alias{MotifRegionLinks}
\title{Link TFs to peaks using motifs}
\usage{
MotifRegionLinks(motif_ix)
}
\arguments{
\item{motif_ix}{Binary motif matrix of regions by TFs specifying which TFs have motifs in which regions}
}
\value{
TF region motif overlaps in dataframe format
}
\description{
Link TFs to peaks using motifs
}
