% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{parse_bed_directory}
\alias{parse_bed_directory}
\title{Parse bed files}
\usage{
parse_bed_directory(dir.path)
}
\arguments{
\item{dir.path}{Path to directory with bed files}
}
\value{
List of genomic ranges, one for each file in the directory
}
\description{
Parse bed files
}
