% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{parse_grasp}
\alias{parse_grasp}
\title{Parse files downloaded from the GRASP database}
\usage{
parse_grasp(file.path)
}
\arguments{
\item{file.path}{Path to sumstats file}
}
\value{
Vector of z-scores for all SNPs in the sumstats file
}
\description{
Parse files downloaded from the GRASP database
}
