% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{parse_ld_blocks}
\alias{parse_ld_blocks}
\title{Parse LD blocks from file}
\usage{
parse_ld_blocks(file.path)
}
\arguments{
\item{file.path}{Path to LD blocks file}
}
\value{
List of snps contained within each LD block
}
\description{
Parse LD blocks from file
}
