% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{parse_sumstats}
\alias{parse_sumstats}
\title{Parse munged sumstats files (same format as input to LDSC)}
\usage{
parse_sumstats(file.path, p.cutoff = 0.001)
}
\arguments{
\item{file.path}{Path to sumstats file}

\item{p.cutoff}{P-value cutoff for keeping a SNP}
}
\value{
Vector of z-scores for all SNPs in the sumstats file
}
\description{
Parse munged sumstats files (same format as input to LDSC)
}
