% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{peak_gene_coaccess}
\alias{peak_gene_coaccess}
\title{Compute relationship between peaks and genes using coaccessibility/promoters}
\usage{
peak_gene_coaccess(
  peaks.gr,
  conns,
  link.promoter = T,
  promoter.region = c(-5000, 5000),
  anno.level = "transcript"
)
}
\arguments{
\item{peaks.gr}{Peaks in granges format}

\item{conns}{Dataframe of peak to peak coaccessibility}

\item{link.promoter}{Include peaks in gene promoters}

\item{promoter.region}{Specify the window around the TSS that counts as a promoter}

\item{anno.level}{Specify "gene" or "transcript" for a gene/transcript level annotation}
}
\value{
Matrix of peak to gene coaccessibilities. A score of 1 means the peak is in the gene promoter.
}
\description{
Compute relationship between peaks and genes using coaccessibility/promoters
}
