% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Visualize a TF - gene network using igraph}
\usage{
PlotNetwork(tf.gene.df, plot.title = NULL, label = F, label.size = 0.5)
}
\arguments{
\item{tf.gene.df}{Dataframe with a TF column and gene column}

\item{plot.title}{Title of the plot}

\item{label}{Labels the TFs/genes in the graph}

\item{label.szie}{Size of TF/gene labels}
}
\value{
Plot visualizing the TF - gene network using a force directed layout
}
\description{
Visualize a TF - gene network using igraph
}
