% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{region_gene_coaccess}
\alias{region_gene_coaccess}
\title{Compute relationship between arbitrary and genes using coaccessibility between peaks/promoters}
\usage{
region_gene_coaccess(
  regions,
  regions.anno,
  conns,
  hg38.chr.lengths,
  link.promoter = F,
  promoter.region = c(-5000, 5000),
  anno.level = "gene",
  buffer.size = 1000,
  region.name = "HAR"
)
}
\arguments{
\item{regions}{Genomic regions in granges format}

\item{regions.anno}{Annnotated genomic regions}

\item{conns}{Dataframe of peak to peak coaccessibility}

\item{hg38.chr.lengths}{Chromosome lengths}

\item{link.promoter}{Include peaks in gene promoters}

\item{promoter.region}{Specify the window around the TSS that counts as a promoter}

\item{anno.level}{Specify "gene" or "transcript" for a gene/transcript level annotation}

\item{buffer.size}{Buffer size around each region}

\item{region.name}{Name of region identifier}
}
\value{
Matrix of region to gene coaccessibilities. A score of 1 means the region is in the gene promoter.
}
\description{
Compute relationship between arbitrary and genes using coaccessibility between peaks/promoters
}
