% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_enrich.R
\name{RegionChromVAR}
\alias{RegionChromVAR}
\title{Run chromVAR on a subset of peaks overlapping a list of regions}
\usage{
RegionChromVAR(SE, motif_ix, regionList, n.runs)
}
\arguments{
\item{SE}{Peaks accessibility matrix in SummarizedExperiment format}

\item{motif_ix}{Transcription factor motif data in SummarizedExperiment format (from motifmatchr)}

\item{n.runs}{Number of replicates to run}

\item{regionsList}{List of regions in GenomicRanges format}
}
\value{
For each set of regions z-scores representing the enrichment of TF motifs in peaks overlapping those regions for each column in SE
}
\description{
Run chromVAR on a subset of peaks overlapping a list of regions
}
