% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{RegionGeneLinks}
\alias{RegionGeneLinks}
\title{Compute relationship between arbitrary regions and genes using either Hi-C or coaccessibility}
\usage{
RegionGeneLinks(
  regions,
  conns,
  link.promoter = T,
  promoter.region = c(-3000, 3000),
  anno.level = "transcript",
  region.name = NULL,
  weight.col = NULL
)
}
\arguments{
\item{regions}{Genomic regions in granges format}

\item{conns}{Dataframe of peak to peak Hi-C/coaccessibility}

\item{link.promoter}{Include peaks in gene promoters}

\item{promoter.region}{Specify the window around the TSS that counts as a promoter}

\item{anno.level}{Specify "gene" or "transcript" for a gene/transcript level annotation}

\item{region.name}{Column name of region identifier. Defaults to peak names}

\item{weight.col}{Column name of weights (i.e. when using coaccessibility). Defaults to 1}
}
\value{
Matrix of region to gene Hi-C/coaccessibility connections
}
\description{
Compute relationship between arbitrary regions and genes using either Hi-C or coaccessibility
}
