% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{RegionOverlapList}
\alias{RegionOverlapList}
\title{Overlap genomic regions}
\usage{
RegionOverlapList(gr1, gr2, gr1.name = "HAR")
}
\arguments{
\item{gr1}{First set of regions to overlap (granges format)}

\item{gr2}{Second set of regions to overlap (granges format)}

\item{gr1.name}{Column with the region names in gr1}
}
\value{
List of overlapping regions
}
\description{
Overlap genomic regions
}
