% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize the first letter of every word in a string}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{String}
}
\value{
String where the first letter of every word is capitalized
}
\description{
Capitalize the first letter of every word in a string
}
