% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{SubsetLinks}
\alias{SubsetLinks}
\title{Subset networks by TF, region, or gene.}
\usage{
SubsetLinks(
  tf.region.df,
  region.gene.df,
  tfs = NULL,
  genes = NULL,
  regions = NULL,
  regions.delim = c(":", "-")
)
}
\arguments{
\item{tf.region.df}{TF - region links}

\item{region.gene.df}{Region - gene links}

\item{tfs}{TFs to keep}

\item{genes}{Genes to keep}

\item{regions}{Genomic regions to keep (as character vector). Will also keep any overlapping regions.}

\item{regions.delim}{Delimiters for genomic regions}
}
\value{
List of dataframes containing TF - region links, region - gene links, and TF - gene links
}
\description{
Subset networks by TF, region, or gene.
}
