% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{TFGeneLinks}
\alias{TFGeneLinks}
\title{Link TFs to genes}
\usage{
TFGeneLinks(tf.region.df, region.gene.df)
}
\arguments{
\item{tf.region.df}{Dataframe linking TFs to regions using motif data}

\item{region.gene.df}{Dataframe linking regions to genes using Hi-C or coaccessibility data}
}
\value{
Dataframe linking TFs to genes
}
\description{
Link TFs to genes
}
