% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{TFRegionLinks}
\alias{TFRegionLinks}
\title{Link TFs to regions using motifs and correlations between TF motif activity and region accessibility}
\usage{
TFRegionLinks(counts, embeddings, regions, bin.size = 50, n.cores = 4)
}
\arguments{
\item{counts}{Single cell accessibility counts matrix (peaks by cells)}

\item{embeddings}{Dimensionality reduction used to generate bins (cells by dimensions)}

\item{regions}{Regions to link TFs to. If null will link TFs to all peaks.}

\item{bin.size}{Number of cells per bin}

\item{n.cores}{Number of cores}
}
\value{
Dataframe of TF to region links and the correlation between TF motif activity and accessibility
}
\description{
Link TFs to regions using motifs and correlations between TF motif activity and region accessibility
}
