% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccess.R
\name{TFRegionMotifs}
\alias{TFRegionMotifs}
\title{Link TFs to regions overlapping peaks with a TF motif binding site}
\usage{
TFRegionMotifs(regions, peaks.gr, motif_ix_mat, region.name = NULL)
}
\arguments{
\item{regions}{Regions (in GRanges format) to link to TFs}

\item{peaks.gr}{Peaks (in GRanges format)}

\item{motif_ix_mat}{Matrix specifying peaks linked to each TF. Rows must match peaks.gr exactly}

\item{region.name}{Name of regions (can be NULL)}
}
\value{
List of regions linked to TFs via binding site motifs
}
\description{
Link TFs to regions overlapping peaks with a TF motif binding site
}
