% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_access.R
\name{top_diff_peaks-deprecated}
\alias{top_diff_peaks-deprecated}
\title{top_diff_peaks}
\arguments{
\item{diff_peaks_list}{List of dA peaks per cluster. Output by find_all_diff}

\item{qval.cutoff}{Q-value cutoff to call a site differentially accessible}

\item{specificity.cutoff}{Specificity score cutoff}

\item{logfc.cutoff}{Log fold-change cutoff}

\item{max.peaks}{Max number of peaks to keep for each cluster}
}
\value{
Filtered list of granges with cluster specificity, logfc, and p-value for each site tested
}
\description{
Subset to top differentially accessible peaks for each cluster
}
\seealso{
\code{\link{chromfunks-deprecated}}
}
\keyword{internal}
