% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_access.R
\name{TopAccessPeaks}
\alias{TopAccessPeaks}
\title{Subset to top accessible peaks for each cluster regardless of whether the peaks are specific to that cluster}
\usage{
TopAccessPeaks(
  peaks_list,
  prop.cutoff = 0.05,
  dev.cutoff = 1,
  max.peaks = 20000
)
}
\arguments{
\item{peaks_list}{List of dA peak info per cluster. The output of CalcDiffAccess.}

\item{prop.cutoff}{Minimum fraction of cells where the peak is accessible}

\item{dev.cutoff}{Minimum deviation from expected accessibility}

\item{max.peaks}{Max number of peaks to keep for each cluster}
}
\value{
Filtered list of granges with proportion accessible, deviation for each site tested
}
\description{
Subset to top accessible peaks for each cluster regardless of whether the peaks are specific to that cluster
}
