% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_access.R
\name{TopDiffPeaks}
\alias{TopDiffPeaks}
\title{Subset to top differentially accessible peaks for each cluster}
\usage{
TopDiffPeaks(
  peaks_list,
  qval.cutoff = 0.05,
  specificity.cutoff = 0,
  logfc.cutoff = 0,
  dev.cutoff = 0,
  min.prop = 0,
  max.peaks = 30000,
  order.by = "qval"
)
}
\arguments{
\item{peaks_list}{List of dA peak info per cluster. The output of CalcDiffAccess.}

\item{qval.cutoff}{Q-value cutoff to call a site differentially accessible}

\item{specificity.cutoff}{Specificity score cutoff}

\item{logfc.cutoff}{Log fold-change cutoff}

\item{dev.cutoff}{Minumum deviation above expected accessibility within the cluster}

\item{min.prop}{Minimum fraction of cells in the cluster with the site accessible}

\item{max.peaks}{Max number of peaks to keep for each cluster}
}
\value{
Filtered list of granges with cluster specificity, logfc, and p-value for each site tested
}
\description{
Subset to top differentially accessible peaks for each cluster
}
