% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{UnflattenVec}
\alias{UnflattenVec}
\title{Unflatten a vector into a matrix}
\usage{
UnflattenVec(v, delim = "_")
}
\arguments{
\item{v}{Vector}

\item{delim}{Delimiter to separate row and column names}
}
\value{
Matrix
}
\description{
Unflatten a vector into a matrix
}
