% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{FlattenMatrix}
\alias{FlattenMatrix}
\title{Reformat rownames}
\usage{
FlattenMatrix(m)
}
\arguments{
\item{m}{Dense matrix}
}
\value{
The same matrix in dataframe format
}
\description{
Reformat rownames
}
