% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{get_snp_data}
\alias{get_snp_data}
\title{Download SNP data from dbSNP using biomaRt}
\usage{
get_snp_data(snps)
}
\arguments{
\item{snps}{Vector of SNP rsIDs}
}
\value{
Dataframe with the genomic location of each SNP (in hg38 coordinates)
}
\description{
Download SNP data from dbSNP using biomaRt
}
