% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_access.R
\name{TopDiffPeaks}
\alias{TopDiffPeaks}
\title{Subset to top differentially accessible peaks for each cluster}
\usage{
TopDiffPeaks(
  peaks_list,
  qval.cutoff = 0.05,
  specificity.cutoff = 1e-04,
  logfc.cutoff = 0.5,
  max.peaks = 20000
)
}
\arguments{
\item{peaks_list}{List of dA peak info per cluster. The output of CalcDiffAccess.}

\item{qval.cutoff}{Q-value cutoff to call a site differentially accessible}

\item{specificity.cutoff}{Specificity score cutoff}

\item{logfc.cutoff}{Log fold-change cutoff}

\item{max.peaks}{Max number of peaks to keep for each cluster}
}
\value{
Filtered list of granges with cluster specificity, logfc, and p-value for each site tested
}
\description{
Subset to top differentially accessible peaks for each cluster
}
